add routes and code preview in demo app
